DHFPROD-6732: Updated based on PR comments
Moved JobDataManager to hub.flow and renamed to JobManager; renamed method to deleteJobs.

Combined DeleteJobsTest and JobDataManagerTest into DeleteJobsTest. 

Removed 7 test batch/job files. The test that used those was verifying that if timeEnded was null or not a valid dateTime, then they weren't deleted. That is now done via a test that nulls out timeEnded on the Job/Batch docs.